LandingPage test case update
landingPageSteps new archive created
