+++ src/test/resources/web/features/logIn.feature
@@ -1,13 +1,13 @@
 @LogIn
 Feature: Clockify log in
 
-  @ToBeAutomated @LogInFromHomePage
+  @ToBeAutomated @LogInFromHomePage @Do
   Scenario: Access to Log in from clockify home page successfully
     Given The client is on landing page
     When The client click on LogIn button
     Then The client get access to clockify LogIn page
 
-  @Smoke @Candidate @LogInManually @Do
+  @Smoke @Candidate @LogInManually
   Scenario: The client manually log in to clockify successfully
     Given The client is on clockify log in page
     When The client click on 'Log in manually'
